+++ Kokoro/Patches/CardPatches.cs
@@ -37,6 +37,7 @@ internal static class CardPatches
 			logger: Instance.Logger!,
 			original: () => AccessTools.DeclaredMethod(typeof(Card), nameof(Card.Render)),
 			prefix: new HarmonyMethod(typeof(CardPatches), nameof(Card_Render_Prefix)),
+			postfix: new HarmonyMethod(typeof(CardPatches), nameof(Card_Render_Postfix)),
 			transpiler: new HarmonyMethod(typeof(CardPatches), nameof(Card_Render_Transpiler))
 		);
 		harmony.TryPatch(
@@ -116,6 +117,39 @@ internal static class CardPatches
 		CurrentResourceState = null;
 		CurrentNonDrawingResourceState = null;
 		CardRenderMatrixStack.Clear();
+	}
+
+	private static void Card_Render_Postfix(Card __instance, G g, Vec? posOverride, State? fakeState, double? overrideWidth)
+	{
+		var state = fakeState ?? g.state;
+		if (state.route is not Combat combat)
+			return;
+		if (!Instance.RedrawStatusManager.IsRedrawPossible(state, combat, __instance))
+			return;
+
+		var position = posOverride ?? __instance.pos;
+		position += new Vec(0.0, __instance.hoverAnim * -2.0 + Mutil.Parabola(__instance.flipAnim) * -10.0 + Mutil.Parabola(Math.Abs(__instance.flopAnim)) * -10.0 * (double)Math.Sign(__instance.flopAnim));
+		position += new Vec(((overrideWidth ?? 59) - 21) / 2.0, 82 - 13 / 2.0 - 0.5);
+		position = position.round();
+
+		var result = SharedArt.ButtonSprite(
+			g,
+			new Rect(position.x, position.y, 19, 13),
+			new UIKey((UK)21370099, __instance.uuid),
+			(Spr)ModEntry.Instance.Content.RedrawButtonSprite.Id!.Value,
+			(Spr)ModEntry.Instance.Content.RedrawButtonOnSprite.Id!.Value,
+			onMouseDown: new MouseDownHandler(() =>
+			{
+				if (Instance.RedrawStatusManager.GetHandlingHook(state, combat, __instance) is not { } hook)
+					return;
+
+				hook.PayForRedraw(state, combat, __instance);
+				hook.DoRedraw(state, combat, __instance);
+				Instance.RedrawStatusManager.AfterRedraw(state, combat, __instance, hook);
+			})
+		);
+		if (result.isHover)
+			g.tooltips.Add(position + new Vec(30, 10), Instance.Api.GetRedrawStatusTooltip());
 	}
 
 	private static IEnumerable<CodeInstruction> Card_Render_Transpiler(IEnumerable<CodeInstruction> instructions, MethodBase originalMethod, ILGenerator il)